+++ examples/data-sources/nxos_physical_interface_vrf/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_physical_interface_vrf" "example" {
+  interface_id = "eth1/10"
+}