+++ examples/resources/nxos_physical_interface_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_physical_interface_vrf.example "sys/intf/phys-[eth1/10]/rtvrfMbr"