+++ examples/resources/nxos_physical_interface_vrf/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_physical_interface_vrf" "example" {
+  interface_id = "eth1/10"
+  vrf_dn       = "sys/inst-default"
+}